Move Slack message model to API, since it's the only place left that uses it
